Parse a duration
